PostgresSqlAction.prepareSession: disable the processing of JDBC parameter metadata
Version 42.2.11 of the PostgreSQL JDBC driver introduced a backwards-incompatible
change in PgDatabaseMetaData.getProcedures, called internally, to only return
procedures (introduced in PSQL 11).

See https://github.com/spring-projects/spring-framework/issues/25399 for more details
As a workaround, we disable the processing of parameter metadata obtained via JDBC and
explicitly redeclare the function parameters.

This change is the least-invasive requirement in order the upgrade the
org.postgresql:postgresql dependency at the omero-server component.